refactor: 패키지 이동 및 ktFormat 적용
